v1.0.1.0
##### ADDED
* VerifyInpatientAdmissionDate.cs
* CompleteOptionObject.cs
##### CHANGED
* AvatoolWebService.VerifyInpatientAdmissionDate() => VerifyInpatientAdmissionDate.VerifyInpatientAdmissionDate()
* AvatoolWebService.CompleteOptionObject() => CompleteOptionObject.CompleteOptionObject()
