+++ src/AppData/Doc/changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## v1.0.1.0 (2019-11-14)
+##### ADDED
+* VerifyInpatientAdmissionDate.cs
+* CompleteOptionObject.cs
+##### CHANGED
+* AvatoolWebService.VerifyInpatientAdmissionDate() => VerifyInpatientAdmissionDate.VerifyInpatientAdmissionDate()
+* AvatoolWebService.CompleteOptionObject() => CompleteOptionObject.CompleteOptionObject()
+
 ## v1.0.0.0 (2019-11-14)
 ##### INFORMATION   
 * Initial release
