+++ src/Avatool-Web-Service.csproj
@@ -75,6 +75,12 @@
       <DependentUpon>AvatoolWebService.asmx</DependentUpon>
       <SubType>Component</SubType>
     </Compile>
+    <Compile Include="CompleteOptionObject.cs">
+      <SubType>Component</SubType>
+    </Compile>
+    <Compile Include="VerifyInpatientAdmissionDate.cs">
+      <SubType>Component</SubType>
+    </Compile>
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
